refactor: extract runner.py as single orchestration layer for agent turns
